🐛 Debias clock leeway on auth tokens (#150)
Previously, we added 2 hours of leeway to the expiration. This is resulting in a significant percentage of our requests failing due to `jwt.exceptions.ImmatureSignatureError: The token is not yet valid (iat)`. This change adjusts the values to be an hour in the past or future, hopefully reducing the incidence of this error.