🧵 Reentrant encrypt/decrypt methods
Leverage Promise to replace existing 'async lazy' fields with promises. They have the 'at most once' evaluation property, so are lazy and lock. This does have the side effect that creating a client immediately gets the KAS public key, which is a small change in behavior, although one that should generally improve things.