validation.ts regex fixes
The `.` character in a regex can match any character.  This needs to be escaped in order to match a literal `.`.

This not only strengthens the patterns, but also will help make the `www` search more efficient.  In addition the WWW_HOST TLD part of the pattern was switched from `^s` to `a-z`.  `^s` will match any character except `s`, it was assumed that `^\s` was expected (to match any non-whitespace character).  However the `a-z` check is simpler.