validation.ts: Further simplify WWW_HOST pattern
This removes the `www` prefix check as that condition will be handled by the following group (which with the `+` can match subdomains).

In addition the subdomain portion was further simplified.  By using a `*` instead of two nested optional groups the refex engine should be more efficent at matching the pattern.