fix: Changes to make regex patterns more efficient, accurate and simpler (#376)
* validation.ts regex fixes

The `.` character in a regex can match any character.  This needs to be escaped in order to match a literal `.`.

This not only strengthens the patterns, but also will help make the `www` search more efficient.  In addition the WWW_HOST TLD part of the pattern was switched from `^s` to `a-z`.  `^s` will match any character except `s`, it was assumed that `^\s` was expected (to match any non-whitespace character).  However the `a-z` check is simpler.

* tdf.ts: Make html payload regex more efficient

The key change in this regex is from `[^>]*` to to `[^>]*?`.  The added `?` will change this to a non-greed match, resulting in as few characters as possible to match (vs as many as possible).  This will reduce the amount of backtracking the regex engine will need to do when evaluating.

* validation.ts: WWW_HOST switched to not capture any groups

This removes the negative `www.` check as it's not believed that this group is checked.  In additional all groups have been switched to non-capturing to allow the pattern to be more efficient.

* validation.ts: Further simplify WWW_HOST pattern

This removes the `www` prefix check as that condition will be handled by the following group (which with the `+` can match subdomains).

In addition the subdomain portion was further simplified.  By using a `*` instead of two nested optional groups the refex engine should be more efficent at matching the pattern.