feat(sdk): Adds OpenTDF.open method (#485)
This method allocates and returns a reader object, instead of directly returning an annotated webstream. This has several advantages, mostly for dealing with base TDFs at the moment:

- allocating is synchronous and does not result in immediate network calls, allowing these objects to be pre-allocated.
- inspection methods, most notably `attributes` and `manifest`, can be executed with loading less than the full file and without a key access server

The attributes method does not work for remote or encrypted policies, and the manifest method doesn't have an implementation for Nano at the moment. Also potentially useful features, like accessing encrypted metadata, still also require doing a full decrypt, as the KAO processing code is still inline with the decrypt method in readStream